Swap glide for coil & remove unused imports

+++ app/src/main/java/alex/example/movies/domain/model/Movie.kt
@@ -14,7 +14,7 @@ data class MovieDomain(
     val popularity: Double,
     val vote_count: Int,
     val video: Boolean,
-    val vote_average: Double
+    val vote_average: Float
 )
 
 open class MoviePageResultDomain(
